@@ -24,60 +24,61 @@
     "test:lib:ci": "npm run test:lib -- --ci --runInBand",
     "postbuild:lib": "npm run build --prefix projects/ngrx-rtk-query",
     "semantic-release": "semantic-release",
-    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points && husky install"
+    "prepare": "husky install",
+    "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.2.6",
-    "@angular/common": "~11.2.6",
-    "@angular/compiler": "~11.2.6",
-    "@angular/core": "~11.2.6",
-    "@angular/forms": "~11.2.6",
-    "@angular/platform-browser": "~11.2.6",
-    "@angular/platform-browser-dynamic": "~11.2.6",
-    "@angular/router": "~11.2.6",
+    "@angular/animations": "~11.2.7",
+    "@angular/common": "~11.2.7",
+    "@angular/compiler": "~11.2.7",
+    "@angular/core": "~11.2.7",
+    "@angular/forms": "~11.2.7",
+    "@angular/platform-browser": "~11.2.7",
+    "@angular/platform-browser-dynamic": "~11.2.7",
+    "@angular/router": "~11.2.7",
     "@ngneat/lib": "^2.0.0",
     "@ngrx/effects": "^11.0.1",
     "@ngrx/router-store": "^11.0.1",
     "@ngrx/store": "^11.0.1",
-    "@reduxjs/toolkit": "^1.5.0",
-    "@rtk-incubator/rtk-query": "https://pkg.csb.dev/rtk-incubator/rtk-query/commit/4771d0a4/@rtk-incubator/rtk-query",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
+    "@reduxjs/toolkit": "^1.5.1",
+    "@rtk-incubator/rtk-query": "https://pkg.csb.dev/rtk-incubator/rtk-query/commit/5599c4ff/@rtk-incubator/rtk-query",
+    "rxjs": "~6.6.7",
+    "tslib": "^2.1.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@angular-builders/jest": "^11.1.1",
-    "@angular-devkit/build-angular": "~0.1102.5",
+    "@angular-devkit/build-angular": "~0.1102.6",
     "@angular-eslint/builder": "^2.0.2",
     "@angular-eslint/eslint-plugin": "^2.0.2",
     "@angular-eslint/eslint-plugin-template": "^2.0.2",
     "@angular-eslint/schematics": "^2.0.2",
     "@angular-eslint/template-parser": "^2.0.2",
-    "@angular/cli": "~11.2.5",
-    "@angular/compiler-cli": "~11.2.6",
+    "@angular/cli": "~11.2.6",
+    "@angular/compiler-cli": "~11.2.7",
     "@commitlint/cli": "^12.0.1",
     "@commitlint/config-conventional": "^12.0.1",
     "@ngrx/store-devtools": "^11.0.1",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
-    "@tailwindcss/forms": "^0.2.1",
-    "@testing-library/angular": "^10.4.1",
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/user-event": "^13.0.8",
-    "@types/jest": "^26.0.21",
+    "@tailwindcss/forms": "^0.3.2",
+    "@testing-library/angular": "^10.5.0",
+    "@testing-library/jest-dom": "^5.11.10",
+    "@testing-library/user-event": "^13.0.16",
+    "@types/jest": "^26.0.22",
     "@types/node": "^12.20.6",
-    "@typescript-eslint/eslint-plugin": "^4.19.0",
-    "@typescript-eslint/parser": "^4.19.0",
+    "@typescript-eslint/eslint-plugin": "^4.20.0",
+    "@typescript-eslint/parser": "^4.20.0",
     "all-contributors-cli": "^6.20.0",
     "angular-cli-ghpages": "^1.0.0-rc.1",
     "cpx": "^1.5.0",
     "cross-env": "^7.0.3",
-    "eslint": "^7.22.0",
+    "eslint": "^7.23.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-import": "2.22.1",
     "eslint-plugin-jest": "^24.3.2",
-    "eslint-plugin-jest-dom": "^3.6.5",
+    "eslint-plugin-jest-dom": "^3.7.0",
     "eslint-plugin-jsdoc": "32.3.0",
     "eslint-plugin-ngrx": "^1.17.3",
     "eslint-plugin-prefer-arrow": "1.2.3",
