@@ -56,21 +56,24 @@ export const angularHooksModule = ({
       api,
       moduleOptions: { useDispatch, useSelector, getState },
     });
-    const metareducer = buildMetaReducer({ api, moduleOptions: { useDispatch, getState } });
-    safeAssign(api, { usePrefetch, metareducer });
+    const metareducer: MetaReducer<any> = buildMetaReducer({ api, moduleOptions: { useDispatch, getState } });
+    safeAssign(api, { usePrefetch });
+    safeAssign(api as any, { metareducer });
 
     return {
       injectEndpoint(endpointName, definition) {
         const anyApi = (api as any) as Api<any, Record<string, any>, string, string, AngularHooksModule>;
 
         if (isQueryDefinition(definition)) {
-          const { useQuery, useQueryState, useQuerySubscription } = buildQueryHooks(endpointName);
+          const { useQuery, useLazyQuery, useQueryState, useQuerySubscription } = buildQueryHooks(endpointName);
           safeAssign(anyApi.endpoints[endpointName], {
             useQuery,
+            useLazyQuery,
             useQueryState,
             useQuerySubscription,
           });
           (api as any)[`use${capitalize(endpointName)}Query`] = useQuery;
+          (api as any)[`useLazy${capitalize(endpointName)}Query`] = useLazyQuery;
         } else if (isMutationDefinition(definition)) {
           const useMutation = buildMutationHook(endpointName);
           safeAssign(anyApi.endpoints[endpointName], {
