@@ -1,20 +1,26 @@
-import { EndpointDefinition, EndpointDefinitions, MutationDefinition, QueryDefinition } from '@rtk-incubator/rtk-query';
-import { MutationHook, UseQuery } from './hooks-types';
+import { EndpointDefinitions, MutationDefinition, QueryDefinition } from '@rtk-incubator/rtk-query';
+import { DefinitionType } from '@rtk-incubator/rtk-query/dist/esm/ts/endpointDefinitions';
+import { MutationHook, UseLazyQuery, UseQuery } from './hooks-types';
 
-export type TS41Hooks<Definitions extends EndpointDefinitions> = {
-  [K in string & keyof Definitions as TS41HookName<K, Definitions[K]>]: Definitions[K] extends QueryDefinition<
-    any,
-    any,
-    any,
-    any
-  >
-    ? UseQuery<Definitions[K]>
-    : Definitions[K] extends MutationDefinition<any, any, any, any>
-    ? MutationHook<Definitions[K]>
-    : never;
-};
-
-type TS41HookName<
-  K extends string,
-  Definition extends EndpointDefinition<any, any, any, any>
-> = `use${Capitalize<K>}${Definition extends QueryDefinition<any, any, any, any> ? 'Query' : 'Mutation'}`;
+export type TS41Hooks<Definitions extends EndpointDefinitions> = keyof Definitions extends infer Keys
+  ? Keys extends string
+    ? Definitions[Keys] extends { type: DefinitionType.query }
+      ? {
+          [K in Keys as `use${Capitalize<K>}Query`]: UseQuery<
+            Extract<Definitions[K], QueryDefinition<any, any, any, any>>
+          >;
+        } &
+          {
+            [K in Keys as `useLazy${Capitalize<K>}Query`]: UseLazyQuery<
+              Extract<Definitions[K], QueryDefinition<any, any, any, any>>
+            >;
+          }
+      : Definitions[Keys] extends { type: DefinitionType.mutation }
+      ? {
+          [K in Keys as `use${Capitalize<K>}Mutation`]: MutationHook<
+            Extract<Definitions[K], MutationDefinition<any, any, any, any>>
+          >;
+        }
+      : never
+    : never
+  : never;
