@@ -22,7 +22,7 @@ export function shallowEqual(objA: any, objB: any): boolean {
     return false;
   }
 
-  // tslint:disable-next-line: prefer-for-of
+  // eslint-disable-next-line @typescript-eslint/prefer-for-of
   for (let i = 0; i < keysA.length; i++) {
     if (!Object.prototype.hasOwnProperty.call(objB, keysA[i]) || !is(objA[keysA[i]], objB[keysA[i]])) {
       return false;
